Refactor into two files
